requirements.txt
google-api-python-client>=2.24.0
google-auth>=2.6.0
google-auth-oauthlib>=0.4.6
google-cloud-bigquery>=2.27.0
